refactor database connection and table creation logic into separate function module
